Replace Macros with {spec, value} version, basic display runs again
