+++ app/display/representation-javafx/src/main/java/org/csstudio/display/builder/representation/javafx/MacrosTable.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015-2016 Oak Ridge National Laboratory.
+ * Copyright (c) 2015-2023 Oak Ridge National Laboratory.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -333,7 +333,7 @@ public class MacrosTable
     public void setMacros(final Macros macros)
     {
         data.clear();
-        macros.forEach((name, value) -> data.add(new MacroItem(name, value)));
+        macros.forEachSpec((name, value) -> data.add(new MacroItem(name, value)));
         // Add empty final row
         data.add(new MacroItem("", ""));
     }
