+++ app/display/representation-javafx/src/test/java/org/csstudio/display/builder/representation/javafx/MacrosDialogDemo.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015-2016 Oak Ridge National Laboratory.
+ * Copyright (c) 2015-2023 Oak Ridge National Laboratory.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -29,6 +29,8 @@ public class MacrosDialogDemo extends ApplicationWrapper
         final Macros macros = new Macros();
         macros.add("S", "Test");
         macros.add("N", "17");
+        macros.add("ANOTHER", "$(PASSED_IN)");
+        macros.add("ONEMORE", "$(HOME)");
         final MacrosDialog dialog = new MacrosDialog(macros, null);
         System.out.println(dialog.showAndWait());
     }
