+++ core/framework/src/test/java/org/phoebus/framework/macros/MacrosUnitTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015-2019 Oak Ridge National Laboratory.
+ * Copyright (c) 2015-2023 Oak Ridge National Laboratory.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -7,14 +7,14 @@
  *******************************************************************************/
 package org.phoebus.framework.macros;
 
-import org.junit.jupiter.api.Test;
-
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
+
+import org.junit.jupiter.api.Test;
 
 /** JUnit test of macro handling
  *  @author Kay Kasemir
@@ -73,21 +73,19 @@ public class MacrosUnitTest
     {
         // Plain  NAME=VALUE with some spaces
         Macros macros = Macros.fromSimpleSpec("A=1,  B = 2");
-        assertThat(macros.getNames().size(), equalTo(2));
-        assertThat(macros.getValue("A"), equalTo("1"));
-        assertThat(macros.getValue("B"), equalTo("2"));
+        assertThat(macros.toString(), equalTo("[A='1',B='2']"));
+
+        // Specifications hold the values as provided without expanding them
+        macros = Macros.fromSimpleSpec("A=a,  AA = $(A)$(A)");
+        assertThat(macros.toString(), equalTo("[A='a',AA='$(A)$(A)']"));
 
         // Quoted value with spaces and comma
         macros = Macros.fromSimpleSpec("MSG = \"Hello, Dolly\" , B=2");
-        assertThat(macros.getNames().size(), equalTo(2));
-        assertThat(macros.getValue("MSG"), equalTo("Hello, Dolly"));
-        assertThat(macros.getValue("B"), equalTo("2"));
+        assertThat(macros.toString(), equalTo("[MSG='Hello, Dolly',B='2']"));
 
         // Value with escaped quote
         macros = Macros.fromSimpleSpec("MSG = \"This is a \\\"Message\\\" .. \" , B=2");
-        assertThat(macros.getNames().size(), equalTo(2));
-        assertThat(macros.getValue("MSG"), equalTo("This is a \"Message\" .. "));
-        assertThat(macros.getValue("B"), equalTo("2"));
+        assertThat(macros.toString(), equalTo("[MSG='This is a \"Message\" .. ',B='2']"));
     }
 
     /** Test basic macro=value
@@ -121,14 +119,67 @@ public class MacrosUnitTest
         assertThat(MacroHandler.replace(macros, "Escaped \\\\$(S)"), equalTo("Escaped \\$(S)"));
     }
 
+    /** Test macro=$(other_macro)
+     *  @throws Exception on error
+     */
+    @Test
+    public void testMacrosInValue() throws Exception
+    {
+        final Macros macros = new Macros();
+        macros.add("A", "a");
+        macros.add("AA", "$(A)$(A)");
+
+        // MacroValues contain the expanded values
+        assertThat(macros.getValue("A"), equalTo("a"));
+        assertThat(macros.getValue("AA"), equalTo("aa"));
+
+        // Swap values
+        macros.add("B", "b");
+        assertThat(macros.getValue("A"), equalTo("a"));
+        assertThat(macros.getValue("B"), equalTo("b"));
+
+        macros.add("X", "$(A)");
+        macros.add("A", "$(B)");
+        macros.add("B", "$(X)");
+        // Values have been swapped
+        assertThat(macros.getValue("A"), equalTo("b"));
+        assertThat(macros.getValue("B"), equalTo("a"));
+        // Specs contain the history of how values were defined
+        assertThat(macros.toString(), equalTo("[A='a',AA='$(A)$(A)',B='b',X='$(A)',A='$(B)',B='$(X)']"));
+        // Values hold each macro with only one, the effective value
+        assertThat(macros.toExpandedString(), equalTo("[A='b',AA='aa',B='a',X='a']"));
+
+        // OK to define macro with what it is (just a special case of known macro)
+        macros.add("A", "a");
+        assertThat(macros.getValue("A"), equalTo("a"));
+        macros.add("A", "$(A)");
+        assertThat(macros.getValue("A"), equalTo("a"));
+
+        // When using unknown macro, it remains unresolved
+        // (with log message that's not tested here)
+        macros.add("A", "$(UNKNOWN)");
+        assertThat(macros.getValue("A"), equalTo("$(UNKNOWN)"));
+    }
+
+    /** Test macros from specs
+     *  @throws Exception on error
+     */
+    @Test
+    public void testMacrosFromSpecs() throws Exception
+    {
+        final Macros macros = Macros.fromSimpleSpec("A=a,B=b,X=$(A),A=$(B),B=$(X)");
+        System.out.println(macros);
+        assertThat(macros.getValue("A"), equalTo("b"));
+        assertThat(macros.getValue("B"), equalTo("a"));
+    }
+
     /** Test special cases
      *  @throws Exception on error
      */
     @Test
     public void testSpecials() throws Exception
     {
-        final MacroValueProvider macros = new Macros();
-        System.out.println(macros);
+        final Macros macros = new Macros();
         assertThat(macros.toString(), equalTo("[]"));
 
         assertThat(MacroHandler.replace(macros, "Plain Text"), equalTo("Plain Text"));
@@ -165,6 +216,7 @@ public class MacrosUnitTest
         assertThat(MacroHandler.replace(macros, "/$(DERIVED)/$(DERIVED)/"), equalTo("/default/default/"));
 
         macros.add("MAIN", "main");
+        macros.add("DERIVED", "$(MAIN=default)");
         System.out.println(macros);
 
         assertThat(MacroHandler.replace(macros, "$(DERIVED)"), equalTo("main"));
@@ -194,7 +246,7 @@ public class MacrosUnitTest
      *  @throws Exception on error
      */
     @Test
-    public void testRecursion()
+    public void testRecursion() throws Exception
     {
         final Macros macros = new Macros();
         macros.add("S", "$(S)");
@@ -210,11 +262,48 @@ public class MacrosUnitTest
 
         try
         {
+            // S is known, so $(S=a) is replaced with the value of S,
+            // but that's $(S), ending in recursion
             MacroHandler.replace(macros, "Recursive $(S=a) default");
         }
         catch (Exception ex)
         {
             assertThat(ex.getMessage(), containsString(/* [Rr] */ "ecursive"));
         }
+    }
+
+    @Test
+    public void testHierarchy() throws Exception
+    {
+        // Macro specs provided by preferences
+        final Macros prefs = new Macros();
+        prefs.add("P1", "p1");
+        prefs.add("P2", "p2");
+
+
+        // Macro specs provided by the launcher
+        final Macros launcher = new Macros();
+        launcher.add("L", "launcher");
+        launcher.add("P2", "p2 replaced by launcher");
+
+        // Macro specs provided by the display
+        final Macros display = new Macros();
+        display.add("T", "display");
+
+        // Macro specs provided by a group inside the display
+        final Macros group = new Macros();
+        group.add("T", "group");
+
+        // Hierarchically expand specs
+        display.expand(prefs, launcher);
+        group.expand(display);
+        System.out.println("Group specs : " + group);
+        System.out.println("Group values: " + group.toExpandedString());
+
+        // MacroValues contain the expanded values
+        assertThat(group.getValue("P1"), equalTo("p1"));
+        assertThat(group.getValue("P2"), equalTo("p2 replaced by launcher"));
+        assertThat(display.getValue("T"), equalTo("display"));
+        assertThat(group.getValue("T"), equalTo("group"));
     }
 }
